LayoutTests/imported/w3c/web-platform-tests/intersection-observer/isIntersecting-threshold-expected.txt


PASS At initial scroll position 
FAIL Scrolled to half way through target element assert_equals: expected true but got false
PASS Scrolled to target element completely off screen 
